# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorPayload",
        "type": "object"
      },
...
      "ApiResponse_ModelListData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelListData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[ModelListData]",
        "type": "object"
      },
      "ApiResponse_ModelRecommendationData_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelRecommendationData"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[ModelRecommendationData]",
        "type": "object"
      },
...
      "ApiResponse_dict_str__object__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "request_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatus"
          }
        },
        "required": [
          "request_id",
          "status"
        ],
        "title": "ApiResponse[dict[str, object]]",
        "type": "object"
      },
      "ApiStatus": {
        "enum": [
          "success",
          "accepted",
          "error"
        ],
        "title": "ApiStatus",
        "type": "string"
      },
...
      "DevicePreference": {
        "enum": [
          "auto",
          "cpu",
          "cuda",
          "mps"
        ],
        "title": "DevicePreference",
        "type": "string"
      },
...
      "ModelCandidate": {
        "additionalProperties": false,
        "properties": {
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "model_id",
          "score"
        ],
        "title": "ModelCandidate",
        "type": "object"
      },
      "ModelFamily": {
        "enum": [
          "unet",
          "yolo_seg",
          "sam2"
        ],
        "title": "ModelFamily",
        "type": "string"
      },
      "ModelListData": {
        "additionalProperties": false,
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelMetadata"
            },
            "title": "Models",
            "type": "array"
          }
        },
        "title": "ModelListData",
        "type": "object"
      },
      "ModelMetadata": {
        "additionalProperties": false,
        "properties": {
          "adapter_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Path"
          },
          "adapter_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Sha256"
          },
          "applicable_materials": {
            "items": {
              "type": "string"
            },
            "title": "Applicable Materials",
            "type": "array"
          },
          "config_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Sha256"
          },
          "default_threshold": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Threshold"
          },
          "family": {
            "$ref": "#/components/schemas/ModelFamily"
          },
          "health_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Health Error"
          },
          "metric_context": {
            "additionalProperties": true,
            "title": "Metric Context",
            "type": "object"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Metrics",
            "type": "object"
          },
          "model_card_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Card Sha256"
          },
          "model_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "postprocess_profile": {
            "title": "Postprocess Profile",
            "type": "string"
          },
          "preprocess_profile": {
            "title": "Preprocess Profile",
            "type": "string"
          },
          "quality_tier": {
            "$ref": "#/components/schemas/QualityTier"
          },
          "status": {
            "$ref": "#/components/schemas/ModelStatus"
          },
          "supports_box_prompt": {
            "title": "Supports Box Prompt",
            "type": "boolean"
          },
          "variant": {
            "$ref": "#/components/schemas/ModelVariant"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "weight_sha256": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Sha256"
          }
        },
        "required": [
          "model_id",
          "family",
          "variant",
          "quality_tier",
          "version",
          "status",
          "supports_box_prompt",
          "preprocess_profile",
          "postprocess_profile"
        ],
        "title": "ModelMetadata",
        "type": "object"
      },
      "ModelRecommendationData": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ModelCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "requires_user_confirmation": {
            "default": true,
            "title": "Requires User Confirmation",
            "type": "boolean"
          }
        },
        "title": "ModelRecommendationData",
        "type": "object"
      },
      "ModelRecommendationRequest": {
        "additionalProperties": false,
        "properties": {
          "device": {
            "$ref": "#/components/schemas/DevicePreference",
            "default": "auto"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "max_gpu_memory_mb": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Gpu Memory Mb"
          },
          "prefer": {
            "default": "accuracy",
            "pattern": "^(speed|balance|accuracy)$",
            "title": "Prefer",
            "type": "string"
          },
          "roi_mode": {
            "$ref": "#/components/schemas/RoiMode"
          },
          "target_profile": {
            "$ref": "#/components/schemas/ModelVariant",
            "default": "general"
          }
        },
        "required": [
          "image_id",
          "roi_mode"
        ],
        "title": "ModelRecommendationRequest",
        "type": "object"
      },
      "ModelStatus": {
        "enum": [
          "ready",
          "loading",
          "unavailable",
          "disabled"
        ],
        "title": "ModelStatus",
        "type": "string"
      },
      "ModelVariant": {
        "enum": [
          "general",
          "small_particle",
          "large_particle",
          "dense_particle",
          "low_contrast"
        ],
        "title": "ModelVariant",
        "type": "string"
      },
...
      "QualityTier": {
        "enum": [
          "fast",
          "balanced",
          "accurate"
        ],
        "title": "QualityTier",
        "type": "string"
      },
...
      "RoiMode": {
        "enum": [
          "full_image",
          "boxes"
        ],
        "title": "RoiMode",
        "type": "string"
      },
...
    }
  },
  "info": {
    "summary": "Traceable SEM nanoparticle analysis and evidence-backed query API",
    "title": "NanoLoop Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/models": {
      "get": {
        "operationId": "listModels",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ModelStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "family",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ModelFamily"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          },
          {
            "in": "query",
            "name": "variant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ModelVariant"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Variant"
            }
          },
          {
            "in": "query",
            "name": "quality_tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QualityTier"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quality Tier"
            }
          },
          {
            "in": "query",
            "name": "material",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Material"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModelListData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "List Models",
        "tags": [
          "models"
        ]
      }
    },
    "/api/v1/models/recommend": {
      "post": {
        "operationId": "recommendModels",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRecommendationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ModelRecommendationData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "State or revision conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Payload too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Unsupported media type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Internal server error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Frozen route awaiting service integration"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__object__"
                }
              }
            },
            "description": "Dependency unavailable"
          }
        },
        "summary": "Recommend Models",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}